requirement.txt
tensorboard
tensorflow
tensorflow-estimator
tensorflow-gpu
matplotlib
keras-applications
keras-preprocessing
opencv-python
django
image
pillow
sqlparse
IPython
keras
scikit-learn
pydot
graphviz
pylint
imbalanced-learn
seaborn
scikit-image
pandas
numpy 
scipy